requirements.txt
apscheduler==3.11.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
idna==3.11
msal==1.34.0
pycparser==3.0
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
